move deserialization to from_config + restructure tests
