+++ keras/src/layers/core/embedding.py
@@ -112,12 +112,7 @@ class Embedding(Layer):
         self.lora_rank = lora_rank
         self.lora_alpha = lora_alpha if lora_alpha is not None else lora_rank
         self.lora_enabled = False
-        if isinstance(quantization_config, dict):
-            self.quantization_config = (
-                serialization_lib.deserialize_keras_object(quantization_config)
-            )
-        else:
-            self.quantization_config = quantization_config
+        self.quantization_config = quantization_config
 
         if weights is not None:
             self.build()
@@ -302,6 +297,16 @@ class Embedding(Layer):
             config["lora_rank"] = self.lora_rank
             config["lora_alpha"] = self.lora_alpha
         return {**base_config, **config}
+
+    @classmethod
+    def from_config(cls, config):
+        config = config.copy()
+        config["quantization_config"] = (
+            serialization_lib.deserialize_keras_object(
+                config.get("quantization_config", None)
+            )
+        )
+        return cls(**config)
 
     def _quantization_mode_error(self, mode):
         return NotImplementedError(
