+++ keras/src/layers/core/embedding_test.py
@@ -625,3 +625,32 @@ class EmbeddingTest(test_case.TestCase):
         layer.load_own_variables(int4_store)
         self.assertAllClose(layer._embeddings, int4_store["0"])
         self.assertAllClose(layer.embeddings_scale, int4_store["1"])
+
+    def test_embedding_int8_custom_quantizer(self):
+        """
+        Test custom quantizer serialization for embedding layer with
+        int8 quantization.
+        """
+        # Setup
+        weight_range = (-50, 50)
+        config = Int8QuantizationConfig(
+            weight_quantizer=AbsMaxQuantizer(axis=-1, value_range=weight_range),
+        )
+
+        # Build & Quantize
+        layer = layers.Embedding(input_dim=100, output_dim=16)
+        layer.build(None)
+        layer.quantize("int8", config=config)
+
+        # Serialize & Deserialize
+        serialized = layer.get_config()
+        new_layer = layers.Embedding.from_config(serialized)
+
+        # Verify
+        self.assertIsInstance(
+            new_layer.quantization_config, Int8QuantizationConfig
+        )
+        quantizer = new_layer.quantization_config.weight_quantizer
+        self.assertIsInstance(quantizer, AbsMaxQuantizer)
+        self.assertEqual(quantizer.axis, (-1,))
+        self.assertAllEqual(quantizer.value_range, weight_range)
