+++ keras/src/quantizers/quantization_config_test.py
@@ -1,7 +1,5 @@
 import os
 
-from absl.testing import parameterized
-
 from keras.src import layers
 from keras.src import models
 from keras.src import saving
@@ -125,224 +123,6 @@ class QuantizationConfigTest(testing.TestCase):
         with self.assertRaisesRegex(ValueError, "output_dtype='int8'"):
             Int4QuantizationConfig(weight_quantizer=q)
 
-
-class CustomQuantizerSerializationTest(
-    testing.TestCase, parameterized.TestCase
-):
-    def _check_quantizer_config(
-        self, quantizer, valid_class, axis, value_range
-    ):
-        self.assertIsInstance(quantizer, valid_class)
-        self.assertEqual(quantizer.axis, axis)
-
-        # Normalize value_range to list
-        if value_range is not None:
-            self.assertAllEqual(quantizer.value_range, value_range)
-
-    def test_dense_int8_custom_quantizer(self):
-        """
-        Test custom quantizer serialization for dense layer.
-        """
-        # Setup
-        weight_range = (-127, 127)
-        act_range = (-5, 5)
-        config = Int8QuantizationConfig(
-            weight_quantizer=AbsMaxQuantizer(axis=0, value_range=weight_range),
-            activation_quantizer=AbsMaxQuantizer(
-                axis=-1, value_range=act_range
-            ),
-        )
-
-        # Build & Quantize
-        layer = layers.Dense(10)
-        layer.build((None, 5))
-        layer.quantize("int8", config=config)
-
-        # Serialize & Deserialize
-        serialized = layer.get_config()
-        new_layer = layers.Dense.from_config(serialized)
-
-        # Verify
-        self.assertIsInstance(
-            new_layer.quantization_config, Int8QuantizationConfig
-        )
-        self._check_quantizer_config(
-            new_layer.quantization_config.weight_quantizer,
-            AbsMaxQuantizer,
-            axis=(0,),
-            value_range=weight_range,
-        )
-        self._check_quantizer_config(
-            new_layer.quantization_config.activation_quantizer,
-            AbsMaxQuantizer,
-            axis=(-1,),
-            value_range=act_range,
-        )
-
-    def test_dense_int8_weight_only_quantizer(self):
-        """
-        Test custom quantizer serialization for dense layer with
-        weight-only quantization.
-        """
-        # Setup
-        config = Int8QuantizationConfig(
-            weight_quantizer=AbsMaxQuantizer(axis=0),
-            activation_quantizer=None,
-        )
-
-        # Build & Quantize
-        layer = layers.Dense(10)
-        layer.build((None, 5))
-        layer.quantize("int8", config=config)
-
-        # Serialize & Deserialize
-        serialized = layer.get_config()
-        new_layer = layers.Dense.from_config(serialized)
-
-        # Verify
-        self.assertIsInstance(
-            new_layer.quantization_config, Int8QuantizationConfig
-        )
-        self.assertIsInstance(
-            new_layer.quantization_config.weight_quantizer, AbsMaxQuantizer
-        )
-        self.assertIsNone(new_layer.quantization_config.activation_quantizer)
-
-    def test_dense_int4_custom_quantizer(self):
-        """
-        Test custom quantizer serialization for dense layer with
-        int4 quantization.
-        """
-        # Setup
-        weight_range = (-8, 7)
-        act_range = (-2, 2)
-        config = Int4QuantizationConfig(
-            weight_quantizer=AbsMaxQuantizer(axis=0, value_range=weight_range),
-            activation_quantizer=AbsMaxQuantizer(
-                axis=-1, value_range=act_range
-            ),
-        )
-
-        # Build & Quantize
-        layer = layers.Dense(10)
-        layer.build((None, 5))
-        layer.quantize("int4", config=config)
-
-        # Serialize & Deserialize
-        serialized = layer.get_config()
-        new_layer = layers.Dense.from_config(serialized)
-
-        # Verify
-        self.assertIsInstance(
-            new_layer.quantization_config, Int4QuantizationConfig
-        )
-        self._check_quantizer_config(
-            new_layer.quantization_config.weight_quantizer,
-            AbsMaxQuantizer,
-            axis=(0,),
-            value_range=weight_range,
-        )
-        self._check_quantizer_config(
-            new_layer.quantization_config.activation_quantizer,
-            AbsMaxQuantizer,
-            axis=(-1,),
-            value_range=act_range,
-        )
-
-    def test_einsum_dense_int8_custom_quantizer(self):
-        """
-        Test custom quantizer serialization for einsum dense layer with
-        int8 quantization.
-        """
-        # Setup
-        weight_range = (-10, 10)
-        config = Int8QuantizationConfig(
-            weight_quantizer=AbsMaxQuantizer(axis=0, value_range=weight_range),
-            activation_quantizer=None,
-        )
-
-        # Build & Quantize
-        layer = layers.EinsumDense("ab,bc->ac", output_shape=10)
-        layer.build((None, 5))
-        layer.quantize("int8", config=config)
-
-        # Serialize & Deserialize
-        serialized = layer.get_config()
-        new_layer = layers.EinsumDense.from_config(serialized)
-
-        # Verify
-        self.assertIsInstance(
-            new_layer.quantization_config, Int8QuantizationConfig
-        )
-        self._check_quantizer_config(
-            new_layer.quantization_config.weight_quantizer,
-            AbsMaxQuantizer,
-            axis=(0,),
-            value_range=weight_range,
-        )
-
-    def test_embedding_int8_custom_quantizer(self):
-        """
-        Test custom quantizer serialization for embedding layer with
-        int8 quantization.
-        """
-        # Setup
-        weight_range = (-50, 50)
-        config = Int8QuantizationConfig(
-            weight_quantizer=AbsMaxQuantizer(axis=-1, value_range=weight_range),
-        )
-
-        # Build & Quantize
-        layer = layers.Embedding(input_dim=100, output_dim=16)
-        layer.build(None)
-        layer.quantize("int8", config=config)
-
-        # Serialize & Deserialize
-        serialized = layer.get_config()
-        new_layer = layers.Embedding.from_config(serialized)
-
-        # Verify
-        self.assertIsInstance(
-            new_layer.quantization_config, Int8QuantizationConfig
-        )
-        self._check_quantizer_config(
-            new_layer.quantization_config.weight_quantizer,
-            AbsMaxQuantizer,
-            axis=(-1,),
-            value_range=weight_range,
-        )
-
-    def test_reversible_embedding_int8_custom_quantizer(self):
-        """
-        Test custom quantizer serialization for reversible embedding layer with
-        int8 quantization.
-        """
-        # Setup
-        weight_range = (-20, 20)
-        config = Int8QuantizationConfig(
-            weight_quantizer=AbsMaxQuantizer(axis=-1, value_range=weight_range),
-        )
-
-        # Build & Quantize
-        layer = layers.ReversibleEmbedding(input_dim=100, output_dim=16)
-        layer.build(None)
-        layer.quantize("int8", config=config)
-
-        # Serialize & Deserialize
-        serialized = layer.get_config()
-        new_layer = layers.ReversibleEmbedding.from_config(serialized)
-
-        # Verify
-        self.assertIsInstance(
-            new_layer.quantization_config, Int8QuantizationConfig
-        )
-        self._check_quantizer_config(
-            new_layer.quantization_config.weight_quantizer,
-            AbsMaxQuantizer,
-            axis=(-1,),
-            value_range=weight_range,
-        )
-
     def test_model_save_and_load(self):
         """
         Test custom quantizer serialization for model save and load.
@@ -374,11 +154,10 @@ class CustomQuantizerSerializationTest(
         self.assertIsInstance(
             loaded_layer.quantization_config, Int8QuantizationConfig
         )
-        self._check_quantizer_config(
-            loaded_layer.quantization_config.weight_quantizer,
-            AbsMaxQuantizer,
-            axis=(0,),
-            value_range=weight_range,
-        )
+
+        quantizer = loaded_layer.quantization_config.weight_quantizer
+        self.assertIsInstance(quantizer, AbsMaxQuantizer)
+        self.assertEqual(quantizer.axis, (0,))
+        self.assertAllEqual(quantizer.value_range, weight_range)
         self.assertIsNone(loaded_layer.quantization_config.activation_quantizer)
         self.assertTrue(loaded_layer._is_quantized)
